Refactored tank/enemy code to prepare for enemy tanks
